requirements.txt
APScheduler==3.9.1
numpy==1.23.3
pandas==1.4.1
Pillow==9.3.0
pyodbc==4.0.32
pyparsing==3.0.7
requests==2.28.1
schedule==1.1.0
SQLAlchemy==1.4.39
streamlit==1.14.1
